backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pytesseract==0.3.10
Pillow==10.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.30
pymysql==1.1.0
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Hypothetical library, replace with actual if available
cursorai 